backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Werkzeug==2.0.1
psycopg2-binary==2.9.1
pika==1.2.0
redis==4.0.2